backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3
openai==1.54.0
pymupdf==1.24.11
python-docx==1.1.2
openpyxl==3.1.5
xlrd==2.0.1
pandas==2.2.3
faiss-cpu==1.9.0
tiktoken==0.8.0
jira==3.8.0
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.4.0
aiofiles==24.1.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
azure-identity==1.19.0
